+++ CHANGELOG.md
@@ -7,7 +7,8 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ### Added
 
-- Add the ability to mock methods that use `#[inline]` or `#[cold]`.
+- Add the ability to mock methods that use `#[inline]` or `#[cold]`, and
+  methods or traits that use `#[must_use]`.
   ([#555](https://github.com/asomers/mockall/pull/555))
 
 ### Changed