+++ CHANGELOG.md
@@ -7,6 +7,10 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ### Added
 
+- Add the ability to use `#[concretize]` with methods that have an `FnMut`
+  argument.
+  ([#595](https://github.com/asomers/mockall/pull/595))
+
 - Add the ability to mock methods that use `#[auto_enum]`, from the
   `auto_enums` crate.  But only for methods that use RPIT; Mockall can't yet
   handle syntax like `-> Result<(), impl T>`