+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
-## [ Unreleased ] - ReleaseDate
+## [ 0.13.0 ] - 2024-07-21
 
 ### Added